Rating Service. Allow fractional ratings.   I've changed the ratings 'score' from an integer to a float as fractional ratings seems like a reasonable idea to me.   This had impact all through the Java layer, the REST layer, the model and the test code.
git-svn-id: https://svn.alfresco.com/repos/alfresco-enterprise/alfresco/HEAD/root@21128 c4b6b30b-aa2e-2d43-bbcb-ca4b014f7261

+++ source/java/org/alfresco/repo/web/scripts/rating/RatingPost.java
@@ -82,7 +82,7 @@ public class RatingPost extends AbstractRatingWebScript
             
             // Range checking of the rating score will be done within the RatingService.
             // So we can just apply the rating.
-            final int rating = json.getInt(RATING);
+            final float rating = (float)json.getDouble(RATING);
             ratingService.applyRating(nodeRefToBeRated, rating, schemeName);
 
             // We'll return the URL to the ratings of the just-rated node.
